react: rename react folder from uppy-react → react
